Link Qt Quick Controls and Layouts for WebAssembly build
Fix Emscripten type casting ambiguity issue in sound index calculation for WebAssembly build

+++ src/sound_player.cpp
@@ -21,7 +21,7 @@ void SoundPlayer::playBrickCollision() const
 
     if (!sounds.empty())
     {
-        const uint idx = QRandomGenerator::global()->bounded(sounds.size());
+        const uint idx = QRandomGenerator::global()->bounded(static_cast<int>(sounds.size()));
         sounds[idx]->play();
     }
 }
